backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.0",
    "numba==0.58.1",
    "numpy==1.24.4",
    "onnxruntime<1.20",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "rembg>=2.0.50",
    "uvicorn>=0.39.0",
]
